Various fixes

+++ src/hotspot/share/runtime/deoptimization.cpp
@@ -2087,7 +2087,7 @@ JRT_ENTRY(void, Deoptimization::uncommon_trap_inner(JavaThread* current, jint tr
     }
 
     // Load class if necessary
-    if (unloaded_class_index >= 0) {
+    if (((int) unloaded_class_index) >= 0) {
       constantPoolHandle constants(current, trap_method->constants());
       fprintf(stderr, "this is index1: %d\n", unloaded_class_index);
       load_class_by_index(constants, unloaded_class_index, THREAD);
